backend/pyproject.toml
[project]
name = "ir-copilot-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.117.1",
    "starlette>=0.47.2",
    "uvicorn[standard]==0.32.0",
    "python-dotenv==1.0.1",
    "openai==2.20.0",
    "supabase==2.9.1",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-multipart==0.0.22",
    "sse-starlette>=2.0.0,<3.0.0",
    "email-validator==2.3.0",
    "websockets>=13.0,<16",
    "langsmith>=0.2.0",
    "docling>=2.0.0",
    "torch>=2.4",
    "torchvision>=0.19",
    "huggingface_hub[hf-xet]>=0.27",
    "httpx==0.27.2",
    "sentence-transformers>=2.2.0",
    "cohere>=5.0.0",
    "tavily-python>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio>=0.18.0",
]

[tool.uv]
package = false
